Optimize caching and layers
Most of the COPY's for image overriding are moved to the builder stage.
Then one COPY copies all the changes over. This removes about 8 layers that rarely ever change.
Moved the dev EXTIVERSE arg to just before it is used
Also optimized the order so most layers are kept cached
refactoring to make everything more understandable